Temporarily switch `ndecrypt-release` to use the stable v0.2.5 binary.
This is done because the upstream developer has deemed v0.3.0 and v0.3.1 to be broken, so I'm rolling back to the stable v0.2.5 version.
Also due to that rollback v0.2.5 wants Dotnet 6, which I believe is EOL. Trying to build that version and using Dotnet 9's runtime makes it simply error.
Instead I opted for using the upstream provided binary and patch it automatically to work on a Nix system without issue.