Add `executables` info to `sabretools-{git,release}` and rename executable.
Makes it so only `Sabretools` gets put in `$out/bin/`, before renaming it so the executable name is all lowercase. I also remembered to fix it in `meta.mainProgram`!
This is personal preference, but I hope that it makes your life easier.

Oh yeah, I also removed a `with lib;` from the `meta` attrset. That's also personal preference, of course.